Rename component as per the code review.
