+++ homeassistant/components/home_plus_control/manifest.json
@@ -1,8 +1,8 @@
 {
-  "domain": "homepluscontrol",
+  "domain": "home_plus_control",
   "name": "Legrand Home+ Control",
   "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/homepluscontrol",
+  "documentation": "https://www.home-assistant.io/integrations/home_plus_control",
   "requirements": [
     "homepluscontrol==0.0.3"
   ],
