+++ homeassistant/components/home_plus_control/strings.json
