+++ homeassistant/components/home_plus_control/switch.py
