+++ homeassistant/components/home_plus_control/translations/en.json
